+++ src/features/input/elements/Button/Button.js
@@ -1,5 +1,5 @@
 import data from './Button.data.json'
-import createCustomizedElement from '../../../shared/utils/createCustomizedElement'
+import createCustomizedElement from '../../../../utils/createCustomizedElement'
 import handleInput from './handleInput'
 
 const Button = createCustomizedElement(data.name, data.attrs, data.children)
