+++ src/features/output/elements/Heading-2/Heading-2.js
@@ -1,5 +1,5 @@
 import data from './Heading-2.data.json'
-import createCustomizedElement from '../../../shared/utils/createCustomizedElement'
+import createCustomizedElement from '../../../../utils/createCustomizedElement'
 
 const Heading2 = createCustomizedElement(data.name, data.attrs, data.children)
