+++ src/features/output/elements/Section/Section.js
@@ -1,4 +1,4 @@
-import createCustomizedElement from '../../../shared/utils/createCustomizedElement'
+import createCustomizedElement from '../../../../utils/createCustomizedElement'
 import data from './Section.data.json'
 
 const Section = createCustomizedElement(data.name, data.attrs, data.children)
